# Join Hash Table Probing Optimization: Optional Probing Selection Vector