Hi dear DuckDB Team, 

This PR suggests a rewrite of the Hash Table Probing, specifically the `GetRowPointers` function in the `JoinHashTable`. 

It builds on the following three findings: 
1. The `SelectionVector` for only probe for specific tuples is only used (1) when filtering `null` values first or (2) when probing for the rows again in a second iteration if there was a collision before. While the first only happens with specific datasets, the second condition will happen only if any tuple had a salt collision, so in 1 - (65535 / 65536 ^ 2048) =  3% of all times and then only for very small amounts of tuples. However, always having a `SelectionVector` is hindering performance as it means extra operations and is suboptimal for SIMD. Therefore, the probing now only has an optional `SelectionVector`. 

2. The way we densified the vector for filtering out non-matching tuples was suboptimal, as now that we know whether we have a `SelectionVector` or not, we can simply call `Flatten,` which will do nothing if the Vector is already flat.

3. Storing easily from the hash computable values like `ht_offset` and `salt` is inefficient as we can compute them in one cycle, which takes the same time as loading them. However, storing them means (1) load hash, (2) compute offset with AND, (3) store them, and (4) load them later when needed. Steps 1-3 would be SIMDized, but still, it is not worth it. 

I ran the new version on a Mac Book Pro with M4 chip and 24 GB of memory. With micro benchmarks, I saw a around 5% probing speed increase for filtering joins and 10% for non-filtering joins. Non-filtering joins benefit more as the extra lookup for non-matching elements is now removed, which was an overhead for this case.

There are also improvements for TPC-H on my Mac book, however, CI/CD Regression tests don't show similar results.. The results are better for low thread numbers and high scale factors as hash table probing becomes more dominant compared to the rest of the operations. The script to create these results can be found [here](https://github.com/cwida/benchmarker/blob/a4dfd9d80ae5b165cdf47b83969a3cc6eb4a6342/experiments/duckdb_optional_probe_sel.py).

Here are the TPC-H average runtimes per query: 

##### Based on System Setting
| System Setting   | DuckDB (Nightly Build Locally) Runtime   | DuckDB (Optional Probe Sel) Runtime   | Percentage Difference   |
|:-----------------|:-----------------------------------------|:--------------------------------------|:------------------------|
| n_threads=1      | 2.1490s                                  | 2.0258s                               | 6.08%                   |
| n_threads=4      | 0.5734s                                  | 0.5454s                               | 5.13%                   |
| n_threads=8      | 0.3181s                                  | 0.3092s                               | 2.88%                   |

##### Based on Data Configuration
| Data Configuration   | DuckDB (Nightly Build Locally) Runtime   | DuckDB (Optional Probe Sel) Runtime   | Percentage Difference   |
|:---------------------|:-----------------------------------------|:--------------------------------------|:------------------------|
| sf=1                 | 0.0330s                                  | 0.0322s                               | 2.28%                   |
| sf=3                 | 0.0958s                                  | 0.0938s                               | 2.05%                   |
| sf=10                | 0.3427s                                  | 0.3332s                               | 2.87%                   |
| sf=30                | 1.0627s                                  | 1.0397s                               | 2.21%                   |
| sf=100               | 4.2042s                                  | 3.9287s                               | 7.01%                   |

Please let me know if you would like more details or something reworked! While this improves performance, it also makes the code more complicated, but I tried to keep it as tidy as possible.. :)